requirements.txt
numpy>=1.21.0
pandas>=1.3.0
dask>=2021.6.0
numba>=0.53.0
cupy-cuda11x>=10.0.0
cudf-cuda11x>=22.12.0
cython>=0.29.24
ray>=1.13.0
plotly>=5.1.0
datashader>=0.13.0
pytest>=6.2.5
black
isort
mypy
flake8
